* actors.cc: Workaround for possible cxx bug causing a segfault in debug msgs * gamedat.cc (read_saveinfo): Typecast pointer to long rather than int (fixes warnings on 64 bit archs, note that this wasn't a real bug). * usecode/useval.cc: #include <cstdio> disabled on ALPHA_LINUX_CXX
The 'cxx bug' thing was essentially a

=-
cout << (foo? bar1 : bar2) << ...
=-

being replaced with the equivalent

=-
type quux = bar2;
if (foo) quux = bar1;
cout << quux << ...
=-

So, IMHO, this indeed worked around YACB.


-- Christoph

+++ exult/trunk/gamedat.cc
@@ -401,7 +401,7 @@ void Game_window::write_saveinfo()
 	out.put(timeinfo->tm_sec);	// 15
 
 	// Packing for the rest of the structure
-	for (j = (int)&(((SaveGame_Details *)0)->reserved0); j < sizeof(SaveGame_Details); j++)
+	for (j = (long)&(((SaveGame_Details *)0)->reserved0); j < sizeof(SaveGame_Details); j++)
 		out.put(0);
 
 	for (i=0; i<party_size ; i++)
@@ -433,7 +433,7 @@ void Game_window::write_saveinfo()
 		Write2(out, npc->get_property(Actor::health));
 
 		// Packing for the rest of the structure
-		for (j = (int)&(((SaveGame_Party *)0)->reserved0); j < sizeof(SaveGame_Party); j++)
+		for (j = (long)&(((SaveGame_Party *)0)->reserved0); j < sizeof(SaveGame_Party); j++)
 			out.put(0);
 	}
 
@@ -475,7 +475,7 @@ void Game_window::read_saveinfo(std::ifstream &in,
 	in.get(details->real_second);	// 15
 
 	// Packing for the rest of the structure
-	for (j = (int)&(((SaveGame_Details *)0)->reserved0); j < sizeof(SaveGame_Details); j++)
+	for (j = (long)&(((SaveGame_Details *)0)->reserved0); j < sizeof(SaveGame_Details); j++)
 		in.get();
 
 	party = new SaveGame_Party[details->party_size];
@@ -500,7 +500,7 @@ void Game_window::read_saveinfo(std::ifstream &in,
 		party[i].health = Read2(in);
 
 		// Packing for the rest of the structure
-		for (j = (int)&(((SaveGame_Party *)0)->reserved0); j < sizeof(SaveGame_Party); j++)
+		for (j = (long)&(((SaveGame_Party *)0)->reserved0); j < sizeof(SaveGame_Party); j++)
 			in.get();
 	}
 }
